fiks_google_auth_retyrn

+++ src/middlewares/authenticate.js
@@ -3,40 +3,35 @@ import { Session } from '../db/models/session.js';
 import { User } from '../db/models/user.js';
 
 export const authenticate = async (req, res, next) => {
-  try {
-
-    const h = req.get('authorization') || req.get('Authorization');
-    let accessToken = h?.startsWith('Bearer ') ? h.split(' ')[1] : null;
-
-    if (!accessToken) {
-      accessToken = req.cookies?.accessToken || null;
-    }
-    if (!accessToken) {
-      res.set('WWW-Authenticate', 'Bearer realm="api", error="missing_token"');
-      throw createHttpError(401, 'No access token');
-    }
-
-    const session = await Session.findOne({ accessToken });
-    if (!session) {
-      res.set('WWW-Authenticate', 'Bearer realm="api", error="invalid_token"');
-      throw createHttpError(401, 'Session not found!');
-    }
-
-    if (new Date(session.accessTokenValidUntil) < new Date()) {
-      await Session.findByIdAndDelete(session._id);
-      res.set('WWW-Authenticate', 'Bearer realm="api", error="invalid_token", error_description="expired"');
-      throw createHttpError(401, 'Access token expired!');
-    }
-
-    const user = await User.findById(session.userId);
-    if (!user) {
-      await Session.findByIdAndDelete(session._id);
-      throw createHttpError(401, 'User not found!');
-    }
-
-    req.user = user;
-    next();
-  } catch (e) {
-    next(e);
+  const authHeader = req.get('Authorization');
+  if (!authHeader) {
+    throw createHttpError(401, 'Please provide auth header');
   }
+
+  const [bearer, accessToken] = authHeader.split(' ');
+
+  if (bearer !== 'Bearer' || !accessToken) {
+    throw createHttpError(401, 'Token should be of type Bearer');
+  }
+
+  const session = await Session.findOne({ accessToken });
+  if (!session) {
+    throw createHttpError(401, 'Session not found!');
+  }
+
+  if (session.accessTokenValidUntil < new Date()) {
+    await Session.findByIdAndDelete(session._id);
+    throw createHttpError(401, 'Access token expired!');
+  }
+
+  const user = await User.findById(session.userId);
+
+  if (!user) {
+    await Session.findByIdAndDelete(session._id);
+    throw createHttpError(401, 'User, associated with session, is not found!');
+  }
+
+  req.user = user;
+
+  next();
 };
